Generate grpc service stubs
To ensure a similar behavior as in c++ the python code generator for
grpc stubs will create two files. One reflects the c++ header in the
form of an abstract class and is always replaced when generating the
SDK. The second one is the concrete child implementation and will only
be generated if the file does not exist yet. This means that new
method stubs are indicated for the developer inside the parent file
so he can easily implement them.

This behavior is not perfect. It would be best to also correctly
update the child definitions, but this requires a bigger refactoring
for the python and c++ grpc interface support generator.