Generate python grpc stubs (#74)
* Add gRPC Service Server Factory for stubs

* Generate grpc service stubs

To ensure a similar behavior as in c++ the python code generator for
grpc stubs will create two files. One reflects the c++ header in the
form of an abstract class and is always replaced when generating the
SDK. The second one is the concrete child implementation and will only
be generated if the file does not exist yet. This means that new
method stubs are indicated for the developer inside the parent file
so he can easily implement them.

This behavior is not perfect. It would be best to also correctly
update the child definitions, but this requires a bigger refactoring
for the python and c++ grpc interface support generator.

* Add integration tests for generated ServiceServerFactory

* Add python documentation

* Update velocitas_lib

* Update velocitas_lib imports

* Simplify ServiceImpl template

* Update README

* Remove hard coded service names

* Remove camel_case occurences

* Simplify ServiceImpl template

* Init git repo for CI

* Fix grpc file naming

* Add Horn Client tests

* Fix integration tests

* Reorder Server / Client tests

* Fix typo for Horn client integration tests

* Use the impl version of the Service for the documentation

* Add launcher apps for testing client / server

* Fix grpc generation naming

* Fix running python test launcher

* Add Copyright

* Improve tests

---------

Co-authored-by: Markus Petke <MP91@users.noreply.github.com>